Fixed Warnings

+++ Hyro/src/Hyro/Core/Window.h
@@ -9,10 +9,10 @@ namespace Hyro {
 
 	struct WindowProps {
 		std::string Title;
-		uint32_t Width, Height;
+		uint16_t Width, Height;
 		uint32_t SampleCount;
 		
-		WindowProps(const std::string& title = "New Window", uint32_t width = 1280, uint32_t height = 720, uint32_t sampleCount = 1)
+		WindowProps(const std::string& title = "New Window", uint16_t width = 1280, uint16_t height = 720, uint32_t sampleCount = 1)
 			: Title(title), Width(width), Height(height), SampleCount(sampleCount)
 		{ }
 
@@ -29,8 +29,8 @@ namespace Hyro {
 
 		virtual void OnUpdate(TimeStep deltaTime) = 0;
 
-		virtual uint32_t GetWidth() = 0;
-		virtual uint32_t GetHeight() = 0;
+		virtual uint16_t GetWidth() = 0;
+		virtual uint16_t GetHeight() = 0;
 
 		virtual void SetEventCallback(const EventCallbackFn& callback) = 0;
 		virtual void SetVSync(bool enabled) = 0;
